Move all module files into a separate module folder
