+++ melatonin_inspector/LatestCompiledAssets/BinaryData6.cpp
