+++ melatonin_inspector/LatestCompiledAssets/input_file_list
