+++ melatonin_inspector/melatonin/components/collapsable_panel.h
